Initial Design Additional Modifications
I separated the formatMovieTrailers into several helper functions. Then I renamed it to more align with when we use it for TV show trailers. It is now trailersTemplate.
